Updated Pieces.js and InitialisePieces.js
