+++ src/pieces/piecePositions.js
@@ -1,0 +1,34 @@
+const initialPositions = {
+    "knights": {
+        "row": {
+            "w": [[7, 1], [7, 6]],
+            "b": [[0, 1], [0, 6]]
+        }
+    },
+    "kings": {
+        "row": {
+            "w": [[7, 4]],
+            "b": [[0, 4]]
+        }
+    },
+    "queens": {
+        "row": {
+            "w": [[7, 3]],
+            "b": [[0, 3]]
+        }
+    },
+    "bishops": {
+        "row": {
+            "w": [[7, 2], [7, 5]],
+            "b": [[0, 2], [0, 5]]
+        }
+    },
+    "rooks": {
+        "row": {
+            "w": [[7, 0], [7, 7]],
+            "b": [[0, 0], [0, 7]]
+        }
+    }
+};
+
+export default initialPositions;
